walloc: fix size type from int => size_t
size_t is used for most moderm implementations of malloc, calloc, etc.
Using int leads to issues, if there are really large chunks of memory used
such as e.g. for 3.2 GB:

Building clunits/ldom index
WALLOC: failed to malloc -1094967296 bytes

+++ include/EST_walloc.h
@@ -38,14 +38,15 @@
 /*=======================================================================*/
 
 #if !defined(__EST_WALLOC_H__)
+#include <stddef.h>
 
 #if defined(__cplusplus)
 extern "C" {
 #endif
 
-void *safe_walloc(int size);
-void *safe_wcalloc(int size);
-void *safe_wrealloc(void *ptr, int size);
+void *safe_walloc(size_t size);
+void *safe_wcalloc(size_t size);
+void *safe_wrealloc(void *ptr, size_t size);
 #define walloc(TYPE,SIZE) ((TYPE *)safe_walloc(sizeof(TYPE)*(SIZE)))
 #define wcalloc(TYPE,SIZE) ((TYPE *)safe_wcalloc(sizeof(TYPE)*(SIZE)))
 #define wrealloc(PTR,TYPE,SIZE) ((TYPE *)safe_wrealloc((void *)(PTR), sizeof(TYPE)*(SIZE)))
